+++ tests/unit/artifact/test_brew.py
@@ -3,7 +3,7 @@ import pytest
 from tmt.steps.prepare.artifact.providers.brew import BrewArtifactProvider
 from tmt.utils import GeneralError
 
-from .utils import (
+from . import (
     MOCK_BUILD_ID,
     MOCK_RPMS_BREW,
     mock_build_api_responses,
